sdl2: support real chip refs #71  svn merge -r 2333:2367 https://amethyst.yui.ne.jp/svn-dev/pc98/np2/branches/yui/support_gimic  svn merge -r 2372:2379 https://amethyst.yui.ne.jp/svn-dev/pc98/np2/branches/yui/support_gimic
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@2381 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ sdl2/win32/misc/tty.h
@@ -16,8 +16,8 @@ public:
 	bool Open(LPCTSTR lpDevName, UINT nSpeed = 0, LPCTSTR lpcszParam = NULL);
 	bool IsOpened() const;
 	void Close();
-	int Read(LPVOID lpcvData, int nDataSize);
-	int Write(LPCVOID lpcvData, int nDataSize);
+	ssize_t Read(LPVOID lpcvData, ssize_t nDataSize);
+	ssize_t Write(LPCVOID lpcvData, ssize_t nDataSize);
 
 private:
 	HANDLE m_hFile;				/*!< ファイル ハンドル */
